apps/backend/app/requirements.txt
fastapi[standard]
pymongo
python-dotenv
python-multipart
bcrypt
cloudinary
PyJWT
pandas
python-magic
motor
boto3
requests